+++ python/ray/rllib/agents/a3c/a3c_torch_policy_graph.py
@@ -10,7 +10,7 @@ import ray
 from ray.rllib.evaluation.postprocessing import compute_advantages, \
     Postprocessing
 from ray.rllib.evaluation.sample_batch import SampleBatch
-from ray.rllib.evaluation.torch_policy_graph_template import build_torch_policy
+from ray.rllib.evaluation.torch_policy_template import build_torch_policy
 
 
 def actor_critic_loss(policy, batch_tensors):
